added
This image is an plot I made in TOPCAT by eyeing the small similarities in the plot of the NGC 2264 cluster stars with the known roAps. You can see the candidates, known, and FOV, where FOV are the stars in the images we collected, analyzing those images to find a roAp candidate.